Move a couple of tests and fix them
